Time calibration
Played around the values of the pulses to find a relation between the time in the code and real time. Found a configuration that gives 1ms for 10 units of time in the code. Modification to the code required, posible good implementation is reducing the time to max1,2 s and shorten the pulse duratons acordingly.